New Search Bar, Font Renderer changes, general fixes, fixed Sounds, Some setting changes and icon rendering and method misplace fix. Also, a general refactor

+++ src/main/java/dev/heliosclient/module/modules/CustomFov.java
@@ -32,7 +32,7 @@ public class CustomFov extends Module_ {
 
     @SubscribeEvent
     public void onTick(TickEvent.CLIENT event) {
-        ((ISimpleOption<Integer>) (Object) mc.options.getFov()).setValueUnrestricted((int) (FOV.value.floatValue()));
+        ((ISimpleOption<Integer>) (Object) mc.options.getFov()).setValueUnrestricted((int) (FOV.value));
     }
 
     @Override
